requirements.txt
streamlit==1.31.0
googlemaps==4.10.0
numpy==1.26.3
matplotlib==3.8.2
folium==0.15.1
streamlit-folium==0.15.1
python-dotenv==1.0.0

